simple async test
assume some backend service that will "do stuff" and a client
that will call the backend's "do stuff" method (which takes time).

- The client does NOT care what the result is, it just
  needs to cause the backend to do its thing.
- The backend uses Sleep to simulate "doing stuff"
- The backend's "do stuff" method does NOT return a value
- The client calls the backend's "do stuff" method as a
  goroutine - so that it doesn't have to wait.

There are no channels - b/c the client doesn't care about the
result of the call to the back end. Even if it did, it might
exit before the backend completes (usually would).